Rename test file to match content
